requirements.txt
gradio
transformers
torch
python-docx
PyPDF2
